// front/src/app/category/model/Loans.ts
export class Loans {
  map: Function;
  id: number;
  nameGame: string;
  nameClient: string;
  loanDate: {
    replace: Function;
    includes: Function;
    substring: Function;
  };
  returnDate: {
    replace: Function;
    concat: Function;
    substring: Function;
    getTime: Function;
  };
}